Comments for candidacies

+++ app/controllers/candidacies_controller.rb
@@ -2,22 +2,12 @@ class CandidaciesController < ApplicationController
 
   include SessionsHelper
 
-  # def index
-  # end
-
-  # def show
-  # end
-
-  # def new
-  # end
-
-  # def create
-  # end
-
-  # def edit
-  # end
-
+  # Update the the candidacy to "accept" a bounty. Only artists that have a
+  # candidacy to the bounty in question may perform this action.
   def update
+    unless signed_in?
+      redirect_to root_path, :error => "You must sign in."
+    end
     @bounty = Bounty.find(params[:id])
     redirect_to root_path, :error => "Only artists may accept bounties." if (currentUser.artist? == false)
     @candidacy = Candidacy.where(:artist_id => currentUser.id, :bounty_id => @bounty.id).first
@@ -26,22 +16,27 @@ class CandidaciesController < ApplicationController
       if @candidacy.update_attributes(params[:candidacy])
         redirect_to root_path, :notice => "Bounty Accepted!"
       else
-        redirect_to edit_bounty_path(@bounty.id), :error => "Error Accepting bounty."
+        redirect_to edit_bounty_path(@bounty.id), :error => "Error accepting bounty."
       end
     else
       redirect_to root_path, :error => "You are not a candidate for that bounty."
     end
   end
 
+  # Remove a candidacy between the currently logged in artist and the specified
+  # bounty.
   def destroy
+    unless signed_in?
+      redirect_to root_path, :error => "You must sign in."
+    end
     @bounty = Bounty.find(params[:id])
     redirect_to root_path, :error => "Only artists may reject bounties." if (currentUser.artist? == false)
     @candidacy = Candidacy.where(:artist_id => currentUser.id, :bounty_id => @bounty.id).first
     unless @candidacy.nil?
       if @candidacy.destroy
-        redirect_to root_path, :notice => "Candidacy Removed!"
+        redirect_to root_path, :notice => "Candidacy removed!"
       else
-        redirect_to edit_bounty_path(@bounty.id), :error => "Error Removing Candidacy."
+        redirect_to edit_bounty_path(@bounty.id), :error => "Error removing candidacy."
       end
     else
       redirect_to root_path, :error => "You are not a candidate for that bounty."
